+++ src/utils/about.py
@@ -1,17 +1,17 @@
 # -*- coding: utf-8 -*-
-__title__ = 'Oracle Tools GUI'
-__package_name__ = 'oracle_tools_gui'
+__title__ = 'Paulissoft Application Tools for Oracle (PATO) GUI'
+__package_name__ = 'pato'
 __author__ = "Gert-Jan Paulissen"
 __description__ = ' '.join('''
-GUI for Oracle Tools.
+Paulissoft Application Tools for Oracle (PATO) GUI.
 '''.strip().split())
 __email__ = "paulissoft@gmail.com"
-__version_info__ = ('1', '2', '0')
+__version_info__ = ('2', '0', '0')
 __version__ = '.'.join(__version_info__)
 __license__ = "MIT License"
-__copyright__ = 'Copyright (c) 2021 Gert-Jan Paulissen'
-__url__ = "https://github.com/paulissoft/oracle-tools-gui"
-__help_url__ = "https://paulissoft.github.io/oracle-tools-gui"
+__copyright__ = 'Copyright (c) 2021-2022 Gert-Jan Paulissen'
+__url__ = "https://github.com/paulissoft/pato-gui"
+__help_url__ = "https://paulissoft.github.io/pato-gui"
 
 if __name__ == '__main__':
     print(__version__)